B-02611: Document getPageKey, pageKey and getPortletWindowKey Changed MenuItemEntity.key to be of type MenuItemKey and removed method getMenuItemKey. Fixed wrong package mapping in hbm files. Moved MenuItemService.getPageKeyByPath to its own class MenuItemKeysByPathResolver.

+++ modules/cms-core/src/main/java/com/enonic/cms/core/content/index/translator/ContentQueryTranslator.java
@@ -140,7 +140,7 @@ public final class ContentQueryTranslator
                 {
                     public void appendValue( StringBuffer sql, MenuItemEntity section )
                     {
-                        sql.append( section.getKey() );
+                        sql.append( section.getKey().toInt() );
                     }
                 };
                 hqlQuery.addFilter( "AND", sectionKeyInClause.toString() );
